+++ changes.md
@@ -1,6 +1,7 @@
 # SYNY CHANGE LOG
 
 ## SYNY-v1.2b
+- `run_syny.pl`: Added the `--min_asize`command line option to filter out alignments/clusters smaller than integer value (e.g. --min_asize 5000)
 - `run_syny.pl`, `get_paf.pl`, `paf_metrics.py`: Added the `--no_sec` command line option to filter out secondary alignments from minimap2 alignments.
 - `gff3_to_gbff.pl`: Fixed a bug that crashed the script when dealing with contigs without annotations.
 - `jgi_to_ncbi_gff.pl`: Added a new script to convert JGI GFF files to a NCBI-like GFF3 format compatible with `gff3_to_gbff.pl`.